pyproject.toml
[project]
name = "docker_actions_example"
version = "0.1.0"
description = ""
authors = [
  {name = "Mihhail Samusev", email = "mhs@bluepp.dk" }
]
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "pydantic==2.*",
    "python-decouple",
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"